requirements.txt
click==8.1.*
pillow==10.0.*
pqdm==0.2.*
tqdm==4.65.*
